tests/requirements.txt
pytest
ragger[speculos]>=1.6.0
ragger[ledgerwallet]>=1.6.0
Jinja2==3.1.2
Flask==2.1.2
ecdsa>=0.16.1,<0.17.0
pysha3>=1.0.0,<2.0.0
